change script

+++ XLM-master/src/model/transformer.py
@@ -131,7 +131,7 @@ class ImgPredLayer(nn.Module):
 
         if self.asm is False:
             scores = self.proj(x).view(-1, self.n_classes)
-            loss = F.cross_entropy(scores, y, reduction='mean')
+            loss = F.cross_entropy(scores, y.long(), reduction='mean')
         else:
             _, loss = self.proj(x, y)
             scores = self.proj.log_prob(x) if get_scores else None
